pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "podcasttf"
version = "0.1.0"
description = "Podcast transcription tool with SRT and AI post-processing"
requires-python = ">=3.10"
dependencies = [
    "requests>=2.31.0",
    "dashscope>=1.20.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
podcasttf = "podcasttf.main:main"
